+++ app/Actions/Search/PhotoSearch.php
@@ -4,7 +4,7 @@ namespace App\Actions\Search;
 
 use App\Contracts\Exceptions\InternalLycheeException;
 use App\DTO\PhotoSortingCriterion;
-use App\Models\Extensions\FixedQueryBuilder;
+use App\Eloquent\FixedQueryBuilder;
 use App\Models\Extensions\SortingDecorator;
 use App\Models\Photo;
 use App\Policies\PhotoQueryPolicy;
