+++ app/Models/BaseAlbumImpl.php
@@ -6,6 +6,7 @@ use App\Constants\AccessPermissionConstants as APC;
 use App\Constants\RandomID;
 use App\Contracts\Models\HasRandomID;
 use App\DTO\PhotoSortingCriterion;
+use App\Eloquent\UseFixedQueryBuilder;
 use App\Enum\ColumnSortingType;
 use App\Enum\OrderSortingType;
 use App\Models\Extensions\HasAttributesPatch;
@@ -13,7 +14,6 @@ use App\Models\Extensions\HasBidirectionalRelationships;
 use App\Models\Extensions\HasRandomIDAndLegacyTimeBasedID;
 use App\Models\Extensions\ThrowsConsistentExceptions;
 use App\Models\Extensions\ToArrayThrowsNotImplemented;
-use App\Models\Extensions\UseFixedQueryBuilder;
 use App\Models\Extensions\UTCBasedTimes;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Collection;
