+++ app/Models/SizeVariant.php
@@ -5,6 +5,7 @@ namespace App\Models;
 use App\Actions\SizeVariant\Delete;
 use App\Casts\MustNotSetCast;
 use App\Contracts\Models\AbstractSizeVariantNamingStrategy;
+use App\Eloquent\UseFixedQueryBuilder;
 use App\Enum\SizeVariantType;
 use App\Exceptions\ConfigurationException;
 use App\Exceptions\MediaFileOperationException;
@@ -14,7 +15,6 @@ use App\Models\Extensions\HasAttributesPatch;
 use App\Models\Extensions\HasBidirectionalRelationships;
 use App\Models\Extensions\ThrowsConsistentExceptions;
 use App\Models\Extensions\ToArrayThrowsNotImplemented;
-use App\Models\Extensions\UseFixedQueryBuilder;
 use App\Models\Extensions\UTCBasedTimes;
 use App\Relations\HasManyBidirectionally;
 use Illuminate\Database\Eloquent\Collection;
