+++ app/Models/User.php
@@ -3,11 +3,11 @@
 namespace App\Models;
 
 use App\Constants\AccessPermissionConstants as APC;
+use App\Eloquent\UseFixedQueryBuilder;
 use App\Exceptions\ModelDBException;
 use App\Exceptions\UnauthenticatedException;
 use App\Models\Extensions\ThrowsConsistentExceptions;
 use App\Models\Extensions\ToArrayThrowsNotImplemented;
-use App\Models\Extensions\UseFixedQueryBuilder;
 use App\Models\Extensions\UTCBasedTimes;
 use Carbon\Exceptions\InvalidFormatException;
 use Illuminate\Database\Eloquent\Collection;
