+++ app/Relations/BaseHasManyPhotos.php
@@ -3,9 +3,9 @@
 namespace App\Relations;
 
 use App\DTO\SortingCriterion;
+use App\Eloquent\FixedQueryBuilder;
 use App\Exceptions\Internal\InvalidOrderDirectionException;
 use App\Models\Extensions\BaseAlbum;
-use App\Models\Extensions\FixedQueryBuilder;
 use App\Models\Extensions\SortingDecorator;
 use App\Models\Photo;
 use App\Policies\PhotoQueryPolicy;
